[#681] refactor: 미들웨어를 관심사별 모듈로 분리
277줄 한 파일에 토큰 갱신, 쿠키 전파, url 판별, 대시보드 인증, 메인 인증,
디바이스 판별이 뒤섞여 있어 어디를 고쳐야 하는지 읽어내기 어려웠다.
동작 변경 없이 배치만 바꾼다.

- session.ts — 쿠키 읽기, 만료 판정, 토큰 갱신, 쿠키 반영, rsc 전파.
  흩어져 있던 session/isExpired/sessionUpdated 지역 변수를 SessionState로 묶어
  가변 상태를 없앤다
- university-url.ts — 학교 코드 판별, 대소문자 정규화, 공개 경로 판별.
  경로가 소문자 학교 코드라는 전제를 이 모듈이 보장한다
- dashboard-auth.ts — 별도 세션을 쓰고 학교 코드 체계를 쓰지 않으므로 분리
- main-auth.ts — 세션 상태와 공개 경로 여부로 통과/로그인을 결정하고
  응답 쿠키를 붙인다

middleware.ts에는 세 단계 조립만 남는다.
