+++ src/middleware.ts
@@ -1,270 +1,18 @@
-import { NextRequest, NextResponse } from 'next/server';
-import {
-  ResponseCookies,
-  RequestCookies,
-} from 'next/dist/compiled/@edge-runtime/cookies';
-import {
-  SESSION_COOKIE_NAME,
-  SESSION_COOKIE_OPTIONS,
-  parseSessionCookie,
-  CookieSession,
-} from '@/shared/lib/cookie-session';
-import {
-  DASHBOARD_SESSION_COOKIE_NAME,
-  parseDashboardSessionCookie,
-} from '@/shared/lib/dashboard-session';
-import getTokenExpiration from '@/shared/lib/getTokenExpiration';
-import { toUrlCode } from '@/shared/lib/urlCodeConverter';
-import { publicRoutes } from '../route';
-
-const API_URL = process.env.NEXT_PUBLIC_API_URL;
-
-/* ────────────────────────────────────────────
- * 미들웨어 전용 유틸리티
- * ──────────────────────────────────────────── */
-
-/**
- * refreshToken을 사용하여 백엔드에 새 accessToken을 요청한다.
- * 미들웨어는 Edge Runtime에서 실행되므로 ky 대신 네이티브 fetch를 사용한다.
- * @returns 새 accessToken 또는 null(실패 시)
- */
-async function refreshAccessToken(
-  refreshToken: string,
-): Promise<string | null> {
-  try {
-    const res = await fetch(`${API_URL}/users/auth/refresh`, {
-      method: 'POST',
-      headers: { Authorization: `Bearer ${refreshToken}` },
-    });
-    if (!res.ok) return null;
-    const data = await res.json();
-    return data.data?.accessToken ?? null;
-  } catch {
-    return null;
-  }
-}
-
-/**
- * 미들웨어에서 response에 설정한 Set-Cookie를 request 헤더에도 반영한다.
- * 이렇게 해야 같은 요청 내 RSC에서 cookies().get()으로 갱신된 값을 읽을 수 있다.
- *
- * 배경: 미들웨어의 response.cookies.set()은 브라우저로 보내는 Set-Cookie 헤더만 설정하고,
- * 원래 요청의 Cookie 헤더는 변경하지 않는다. Next.js는 x-middleware-request-* 헤더를 통해
- * 요청 헤더를 오버라이드할 수 있으므로, 이를 활용하여 RSC에 갱신된 쿠키를 전달한다.
- *
- * @see https://github.com/vercel/next.js/issues/49442
- */
-function applySetCookie(req: NextRequest, res: NextResponse): void {
-  const setCookies = new ResponseCookies(res.headers);
-  const newReqHeaders = new Headers(req.headers);
-  const newReqCookies = new RequestCookies(newReqHeaders);
-
-  setCookies.getAll().forEach((cookie) => newReqCookies.set(cookie));
-
-  NextResponse.next({
-    request: { headers: newReqHeaders },
-  }).headers.forEach((value, key) => {
-    if (
-      key === 'x-middleware-override-headers' ||
-      key.startsWith('x-middleware-request-')
-    ) {
-      res.headers.set(key, value);
-    }
-  });
-}
-
-const NON_UNIVERSITY_PREFIXES = new Set(['admin', 'api', 'test', '_next']);
-const DEFAULT_UNIVERSITY_CODE = 'sejong';
-
-/**
- * 첫 세그먼트가 학교 코드 자리면 그대로 반환한다. 아니면 null.
- * 대문자도 인식해야 정규화 리다이렉트로 넘길 수 있어 대소문자를 가리지 않는다.
- */
-function getUniversitySegment(pathname: string): string | null {
-  const firstSegment = pathname.split('/')[1];
-  if (
-    firstSegment &&
-    !NON_UNIVERSITY_PREFIXES.has(toUrlCode(firstSegment)) &&
-    /^[a-zA-Z]+$/.test(firstSegment)
-  ) {
-    return firstSegment;
-  }
-  return null;
-}
-
-/**
- * 대문자 학교 코드로 들어오면 소문자 경로로 리다이렉트한다.
- * 정규화하지 않으면 공개 경로 판별이 실패해 /sejong/login으로 튕긴다.
- * 이미 소문자면 null을 반환하므로 리다이렉트 루프는 없다.
- */
-function normalizeUniversityCodeCasing(req: NextRequest): NextResponse | null {
-  const segment = getUniversitySegment(req.nextUrl.pathname);
-  if (!segment || segment === toUrlCode(segment)) return null;
-
-  const url = req.nextUrl.clone();
-  const segments = url.pathname.split('/');
-  segments[1] = toUrlCode(segment);
-  url.pathname = segments.join('/');
-
-  return NextResponse.redirect(url);
-}
-
-function stripUniversityCodePrefix(pathname: string): string {
-  if (!getUniversitySegment(pathname)) return pathname;
-  return `/${pathname.split('/').slice(2).join('/')}` || '/';
-}
-
-function extractUniversityCode(pathname: string): string {
-  const segment = getUniversitySegment(pathname);
-  return segment ? toUrlCode(segment) : DEFAULT_UNIVERSITY_CODE;
-}
-
-/**
- * route.ts에 정의된 publicRoutes와 대조하여 인증 없이 접근 가능한 경로인지 판별한다.
- * '/:path*'로 끝나는 패턴은 하위 경로까지 포함한다.
- * /{universityCode}/... 형태의 경로는 universityCode를 제거한 뒤 대조한다.
- */
-function isPublicPath(pathname: string) {
-  const normalizedPath = stripUniversityCodePrefix(pathname);
-  return publicRoutes.some((route) => {
-    if (route.endsWith('/:path*')) {
-      return normalizedPath.startsWith(route.replace('/:path*', ''));
-    }
-    return route === normalizedPath;
-  });
-}
-
-/* ────────────────────────────────────────────
- * 대시보드 전용 인증 처리
- * 메인 서비스 세션(app-session)과 분리된 별도 쿠키(dashboard-session)로 관리한다.
- * ──────────────────────────────────────────── */
-
-function handleDashboardAuth(req: NextRequest): NextResponse | null {
-  const { pathname } = req.nextUrl;
-
-  if (pathname !== '/dashboard' && !pathname.startsWith('/dashboard/'))
-    return null;
-
-  const raw = req.cookies.get(DASHBOARD_SESSION_COOKIE_NAME)?.value;
-  const isAuthenticated = !!parseDashboardSessionCookie(raw)?.accessToken;
-
-  if (pathname === '/dashboard/login') {
-    return isAuthenticated
-      ? NextResponse.redirect(new URL('/dashboard', req.nextUrl))
-      : NextResponse.next();
-  }
-
-  return isAuthenticated
-    ? NextResponse.next()
-    : NextResponse.redirect(new URL('/dashboard/login', req.nextUrl));
-}
+import { NextRequest } from 'next/server';
+import handleDashboardAuth from '@/shared/lib/middleware/dashboard-auth';
+import handleMainAuth from '@/shared/lib/middleware/main-auth';
+import { normalizeUniversityCodeCasing } from '@/shared/lib/middleware/university-url';
 
 export default async function middleware(req: NextRequest) {
-  // 대시보드 경로는 별도 세션으로 처리
+  // 대시보드는 학교 코드 체계도 세션도 별개라 가장 먼저 분리한다
   const dashboardResponse = handleDashboardAuth(req);
   if (dashboardResponse) return dashboardResponse;
 
-  // 아래 경로 판별이 소문자 코드를 전제하므로 세션 처리 전에 정규화한다
+  // 이후 경로 판별이 소문자 코드를 전제하므로 세션 처리 전에 정규화한다
   const casingRedirect = normalizeUniversityCodeCasing(req);
   if (casingRedirect) return casingRedirect;
 
-  const { nextUrl } = req;
-  const sessionCookie = req.cookies.get(SESSION_COOKIE_NAME)?.value;
-  let session: CookieSession | null = parseSessionCookie(sessionCookie);
-  let sessionUpdated = false;
-
-  // ── 1) 토큰 만료 여부 확인 ──
-  // expiresAt이 없으면 만료 판단 불가 → 만료되지 않은 것으로 간주
-  const isExpired = session?.expiresAt
-    ? Date.now() >= session.expiresAt
-    : false;
-
-  // ── 2) 토큰 갱신 ──
-  // accessToken이 만료되었고 refreshToken이 존재하면 갱신을 시도한다.
-  // 갱신 성공 시 세션을 새 토큰으로 교체하고, 실패 시 세션을 null로 만들어 로그아웃 처리한다.
-  if (session && isExpired && session.refreshToken) {
-    const newToken = await refreshAccessToken(session.refreshToken);
-    if (newToken) {
-      session = {
-        ...session,
-        accessToken: newToken,
-        expiresAt: getTokenExpiration(newToken) ?? undefined,
-      };
-      sessionUpdated = true;
-    } else {
-      // refreshToken마저 만료/무효 → 세션 폐기
-      session = null;
-      sessionUpdated = true;
-    }
-  }
-
-  // ── 3) 라우팅 결정 ──
-  if (nextUrl.pathname === '/') {
-    return NextResponse.redirect(
-      new URL(`/${DEFAULT_UNIVERSITY_CODE}`, nextUrl),
-    );
-  }
-
-  const isNowLoggedIn = !!session?.accessToken;
-  const isPublic = isPublicPath(nextUrl.pathname);
-  const universityCode = extractUniversityCode(nextUrl.pathname);
-  const loginUrl = `/${universityCode}/login`;
-
-  let response: NextResponse;
-
-  if (session && isExpired && !session.refreshToken && !sessionUpdated) {
-    // accessToken 만료 + refreshToken 없음 → 세션 쿠키 삭제, 비공개 경로면 로그인으로 리다이렉트
-    response = isPublic
-      ? NextResponse.next()
-      : NextResponse.redirect(new URL(loginUrl, nextUrl));
-    response.cookies.delete(SESSION_COOKIE_NAME);
-  } else if (!isNowLoggedIn && !isPublic) {
-    // 비로그인 상태에서 비공개 경로 접근 → 로그인으로 리다이렉트
-    response = NextResponse.redirect(new URL(loginUrl, nextUrl));
-  } else {
-    // 그 외 → 정상 통과
-    response = NextResponse.next();
-  }
-
-  // ── 4) 갱신된 세션 쿠키 반영 ──
-  // response.cookies.set()은 브라우저에 Set-Cookie 헤더를 전달한다.
-  // RSC로의 전파는 6단계 applySetCookie에서 처리한다.
-  if (sessionUpdated) {
-    if (session) {
-      const cookieValue = encodeURIComponent(JSON.stringify(session));
-      response.cookies.set(
-        SESSION_COOKIE_NAME,
-        cookieValue,
-        SESSION_COOKIE_OPTIONS,
-      );
-    } else {
-      response.cookies.delete(SESSION_COOKIE_NAME);
-    }
-  }
-
-  // ── 5) 디바이스 타입 쿠키 ──
-  // SSR 시점에 User-Agent 기반으로 모바일/데스크탑을 판별하여 쿠키에 저장한다.
-  // 서버 컴포넌트에서 이 쿠키를 읽어 반응형 초기 렌더링에 활용할 수 있다.
-  const userAgent = req.headers.get('user-agent') || '';
-  const isMobile =
-    /Android|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(userAgent);
-  const currentDeviceCookie = req.cookies.get('x-device-type')?.value;
-  const targetDevice = isMobile ? 'mobile' : 'desktop';
-
-  if (currentDeviceCookie !== targetDevice) {
-    response.cookies.set('x-device-type', targetDevice, {
-      path: '/',
-      maxAge: 60 * 60 * 24 * 7, // 7일
-      sameSite: 'lax',
-    });
-  }
-
-  // ── 6) 요청 쿠키 오버라이드 ──
-  // response에 설정한 Set-Cookie를 request 헤더에도 반영하여
-  // 이후 RSC에서 cookies().get()으로 갱신된 값을 읽을 수 있게 한다.
-  applySetCookie(req, response);
-
-  return response;
+  return handleMainAuth(req);
 }
 
 // 정적 자산은 인증과 무관하므로 제외한다. 특히 폰트는 페이지당 수십 개가
